make localAddress and remoteAddress testable in EmbeddedChannel
Motivation:
Testing channel handlers that rely on `Channel.localAddress` or `Channel.remoteAddress` is currently quite inconvenient as one cannot set test values for `localeAddress` or `remoteAddress` on `EmbeddedChannel`.

Modifications:
Change `localAddress` and `remoteAddress` to variables instead of constants and set the address values in `bind()` and `connect()` as proposed in #829.

Result:
Arbitrary values for `localAddress` or `remoteAddress` can be set on `EmbeddedChannel`.