B2MD: check that EOF events are forwarded
Motivation:

Basically I just wasn't 100% sure that I implemented that the EOF events
are not only handled but also forwarded to the next handler in the
pipeline. Quite important.

Modifications:

To figure out if I had implemented that or not I just wrote a test which
this PR is adding.

Result:

One more test, nothing else. Turns out I had implemented that correctly.