Add a HTTP1 client example. It is only a simple echo which closely resembles the TCP and UDP echo examples.
Motivation:
To provide a simple example of the HTTP1 client pipeline.
To create a comparable example to the TCP and UDP demos to show what extra steps are needed for a HTTP layer.

Modifications:
Adds a NIOHTTP1Client example in a main file.
Adds a readme to the NIOHTTP1Client example.

Result:
There is now a simple example of how to send a HTTP1 client request to the server.