Initialize ByteBuffer from ByteBufferView
Motivation:

As Johannes previously said in (#946), there's no reason not be able to
get a ByteBuffer from a ByteBufferView.

Modifications:

I added a public initializer to ByteBuffer taking a ByteBufferView to
create a slice of the view's buffer from the same range.

Result:

`ByteBuffer` will have a public initializer to create a ByteBuffer from
a ByteBufferView.