Provide CircularBuffer.modify(_:_:). (#1009)
Motivation:

When using CircularBuffer as a storage object it would be helpful to be
able to write directly into the underlying storage. As the _modify
accessor is currently underscored and not really safe for us to use, the
next best option is to write a function that takes advantage of the
_modify access on the underlying Array. This patch provides that
function.

Modifications:

- Implemented CircularBuffer.modify(_:_:)

Result:

It'll be possible to use a fast function to access the underlying Array
storage.