Make EmbeddedChannel.isWritable mutable.
Motivation:

When testing programs that check flow control state, it's useful to be
able to control the results of that check. However, right now it's not
possible to set the flow control state of an EmbeddedChannel: it's just
always true. That's not super helpful.

Modifications:

Make EmbeddedChannel.isWritable a stored property that can be changed.

Result:

Better testing.