Give EventLoopFuture a Combine Publisher.
Motivation:

Combine provides users with a nunber of powerful tools for asynchronous
programming that allow users to define a reactive style of programming.
Over time SwiftNIO will want to add a number of hooks for working with
Combine to help enable users to integrate SwiftNIO with their
Combine-using applications.

Modifications:

- Gave EventLoopFuture a Combine Publisher that will provide the result
    of the EventLoopFuture

Result:

Beginnings of Combine integration for SwiftNIO