Updated the EmbeddedChannel tests to check write to the input and output multiple times.
Motivation:
The current test check that the buffer can be written to and then that it is cleared. They do not check that once the buffer is cleared, it can be written to again.

Modifications:
Added a test for both the inbound and outbound buffer to check multiple writes to each buffer.

Result:
We can not be sure that multiple writes to either the inbound or outbound buffers on the embedded channel works.Updated the embedded channel tests to write to the input and output buffers multiple times.