EventLoopPromise.completeWith(Future), mirroring ELF.cascade(to:) #1052
Motivation:

EventLoopFuture and EventLoopPromise are symmetric APIs and for easier
discoverability of being able to complete a promise with a future it
is helpful to have a mirror of future.cascade put onto promise.

Modifications:

- Adds ELP.completeWith(ELF)

Result:

It is now possible to complete a promise with a future,
and not only cascade a future to a promise.