Update the README file to include some recently added examples.
Motivation:

To give visibility to recently added examples.
To satisfy issue #465

Modifications:

Added links to examples in readme.
Slightly re-arranged the order so that echo examples are next to each other in the list.

Result:

You can now see the UDP, HTTP and WebSocket client examples from the README file.