Add API for modifying BB without CoW (#1173)
Motivation:

From time to time a NIO application holds a ByteBuffer that it may want
to edit. Whether it wants to do that may potentially be contingent on
whether having to do so will cause an allocation, as it may be
preferable to prefer a different operation.

Modifications:

- Added API for modifying a ByteBuffer if it will not CoW.

Result:

People can write allocation-sensitive code on ByteBuffer.