Add flatSubmit method to EventLoop (#426) (#1174)
Motivation:

Calling EventLoop's submit method with a task that returns an EventLoopFuture results in a double future. flatSubmit provides an alternative which does not return a double future.

Modifications:

Added flatSubmit() to the EventLoop protocol and provided a default implementation that returns the result of calling flatMap { $0 } on the EventLoopFuture returned from task.

Result:

Additive change only. flatSubmit method on EventLoop will be available for use.