Create ManagedBufferPointer-based FastAtomic class (#1039)
Motivation:

The existing Atomic class holds an UnsafeEmbeddedAtomic which holds an OpaquePointer to raw memory for the C atomic. This results in multiple allocations on init. The new FastAtomic class uses ManagedBufferPointer which tail allocates memory for the C atomic as part of the FastAtomic class allocation.

Modifications:

Created FastAtomic class that uses ManagedBufferPointer to allocate memory for the C atomic. Created version of catmc_atomic_* C functions that expect memory to be allocated/deallocated by caller. Created FastAtomicPrimitive protocol for the new version of catmc_atomic_* functions.

Result:

Purely additive. FastAtomic is available as an alternative to Atomic which results in fewer memory allocations.