close kqueue/epoll fds in Selector.close
Motivation:

There was always this one big embarrassment in the NIO codebase: We
violated the iron rule of 'no fd or scarce resource management in deinit'
in just one place: The most long-lived thing there is in NIO: The
Selector for the SelectableEventLoopGroup. Sure enough, even that tiny
violation lead to issues.

Modifications:

Close the fds in `close` rather than `deinit`.

Result:

- I'm feeling better
- makes @ktoso happy
- fixes #1302